Add backend for unicode implementation
This is first step toward deprecating non-native unicode implementation

ref:

- https://github.com/rails/rails/pull/26743#issuecomment-279277662
- https://github.com/rails/rails/pull/28067#pullrequestreview-23875512
